requirements.txt
Flask
requests
Pillow
numpy